Move generateGlobPatterns to separate module
